+++ fbw-a380x/src/systems/instruments/src/MFD/pages/ATCCOM/MfdAtccomDAtisReceived.tsx
@@ -85,9 +85,6 @@ export class MfdAtccomDAtisReceived extends DisplayComponent<MfdAtccomDAtisRecei
             />
           </div>
         </div>
-        <div style="position: absolute; top: 36px; opacity: 0.3; visibility: hidden;">
-          <img src="/Images/fbw-a380x/d-atis-received.png" alt="" />
-        </div>
         <Footer bus={this.props.bus} mfd={this.props.mfd} fmcService={this.props.fmcService} />
       </>
     );